voice_assistant/requirements.txt
speechrecognition
pyaudio
google-generativeai
pyttsx3
reportlab
config
python-docx
python-pptx
keyring
